Fix not updating error message when authentication fails
If attempt to configure the first network fails, AccessPoints fails to
update error message because it only checks the error code if it is not
active.

With this change, user may see saved network message for a second
because SupplicantStateTracker retries on authentication failure so
intermediate status is displayed between each trial.

Bug: 8284024
Change-Id: I8b976b03878e27e46726ee8a176f131115b7a409